move filterIceServers to util module
pending removal after it has been completly moved to adapter.js
