+++ test/rtcpeerconnection.js
@@ -3427,102 +3427,6 @@ describe('Edge shim', () => {
     });
   });
 
-  describe('filtering of STUN and TURN servers', () => {
-    let pc;
-
-    it('converts legacy url member to urls', () => {
-      pc = new RTCPeerConnection({
-        iceServers: [{url: 'stun:stun.l.google.com'}]
-      });
-      const config = pc.getConfiguration();
-      expect(config.iceServers).to.deep.equal([
-        {urls: 'stun:stun.l.google.com'}
-      ]);
-    });
-
-    it('filters STUN before r14393', () => {
-      RTCPeerConnection = shimPeerConnection(window, 14392);
-      pc = new RTCPeerConnection({
-        iceServers: [{urls: 'stun:stun.l.google.com'}]
-      });
-      const config = pc.getConfiguration();
-      expect(config.iceServers).to.deep.equal([]);
-    });
-
-    it('does not filter STUN without protocol after r14393', () => {
-      pc = new RTCPeerConnection({
-        iceServers: [{urls: 'stun:stun.l.google.com'}]
-      });
-      const config = pc.getConfiguration();
-      expect(config.iceServers).to.deep.equal([
-        {urls: 'stun:stun.l.google.com'}
-      ]);
-    });
-
-    it('does filter STUN with protocol even after r14393', () => {
-      pc = new RTCPeerConnection({
-        iceServers: [{urls: 'stun:stun.l.google.com:19302?transport=udp'}]
-      });
-      const config = pc.getConfiguration();
-      expect(config.iceServers).to.deep.equal([]);
-    });
-
-    it('filters incomplete TURN urls', () => {
-      pc = new RTCPeerConnection({
-        iceServers: [
-          {urls: 'turn:stun.l.google.com'},
-          {urls: 'turn:stun.l.google.com:19302'}
-        ]
-      });
-      const config = pc.getConfiguration();
-      expect(config.iceServers).to.deep.equal([]);
-    });
-
-    it('filters TURN TCP', () => {
-      pc = new RTCPeerConnection({
-        iceServers: [
-          {urls: 'turn:stun.l.google.com:19302?transport=tcp'}
-        ]
-      });
-      const config = pc.getConfiguration();
-      expect(config.iceServers).to.deep.equal([]);
-    });
-
-    describe('removes all but the first server of a type', () => {
-      it('in separate entries', () => {
-        pc = new RTCPeerConnection({
-          iceServers: [
-            {urls: 'stun:stun.l.google.com'},
-            {urls: 'turn:stun.l.google.com:19301?transport=udp'},
-            {urls: 'turn:stun.l.google.com:19302?transport=udp'}
-          ]
-        });
-        const config = pc.getConfiguration();
-        expect(config.iceServers).to.deep.equal([
-          {urls: 'stun:stun.l.google.com'},
-          {urls: 'turn:stun.l.google.com:19301?transport=udp'}
-        ]);
-      });
-
-      it('in urls entries', () => {
-        pc = new RTCPeerConnection({
-          iceServers: [
-            {urls: 'stun:stun.l.google.com'},
-            {urls: [
-              'turn:stun.l.google.com:19301?transport=udp',
-              'turn:stun.l.google.com:19302?transport=udp'
-            ]}
-          ]
-        });
-        const config = pc.getConfiguration();
-        expect(config.iceServers).to.deep.equal([
-          {urls: 'stun:stun.l.google.com'},
-          {urls: ['turn:stun.l.google.com:19301?transport=udp']}
-        ]);
-      });
-    });
-  });
-
   describe('getStats', () => {
     let pc;
     const sdp = SDP_BOILERPLATE + MINIMAL_AUDIO_MLINE;
