+++ util.js
@@ -21,5 +21,43 @@ module.exports = {
       OperationError: undefined
     }[name];
     return e;
+  },
+  // Edge does not like
+  // 1) stun: filtered after 14393 unless ?transport=udp is present
+  // 2) turn: that does not have all of turn:host:port?transport=udp
+  // 3) turn: with ipv6 addresses
+  // 4) turn: occurring muliple times
+  filterIceServers: function(iceServers, edgeVersion) {
+    var hasTurn = false;
+    iceServers = JSON.parse(JSON.stringify(iceServers));
+    return iceServers.filter(function(server) {
+      if (server && (server.urls || server.url)) {
+        var urls = server.urls || server.url;
+        if (server.url && !server.urls) {
+          console.warn('RTCIceServer.url is deprecated! Use urls instead.');
+        }
+        var isString = typeof urls === 'string';
+        if (isString) {
+          urls = [urls];
+        }
+        urls = urls.filter(function(url) {
+          var validTurn = url.indexOf('turn:') === 0 &&
+              url.indexOf('transport=udp') !== -1 &&
+              url.indexOf('turn:[') === -1 &&
+              !hasTurn;
+
+          if (validTurn) {
+            hasTurn = true;
+            return true;
+          }
+          return url.indexOf('stun:') === 0 && edgeVersion >= 14393 &&
+              url.indexOf('?transport=udp') === -1;
+        });
+
+        delete server.url;
+        server.urls = isString ? urls[0] : urls;
+        return !!urls.length;
+      }
+    });
   }
 };
